+++ src/main/java/ar/edu/utn/frbb/tup/model/Profesor.java
@@ -5,7 +5,7 @@ import java.util.Random;
 
 public class Profesor {
 
-    private final long id = 12;
+    private final int id = 12;
     private final String nombre;
     private final String apellido;
     private final String titulo;
